# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuotePriority = typeof QuotePriority[keyof typeof QuotePriority];


export const QuotePriority = {
  low: 'low',
  normal: 'normal',
  urgent: 'urgent',
} as const;

export type QuoteStatus = typeof QuoteStatus[keyof typeof QuoteStatus];


export const QuoteStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Quote {
  id: number;
  name: string;
  email: string;
  phone: string;
  address: string;
  serviceType: string;
  description: string;
  priority: QuotePriority;
  status: QuoteStatus;
  /** @nullable */
  adminNotes?: string | null;
  imageUrls?: string[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type QuoteInputPriority = typeof QuoteInputPriority[keyof typeof QuoteInputPriority];


export const QuoteInputPriority = {
  low: 'low',
  normal: 'normal',
  urgent: 'urgent',
} as const;

export interface QuoteInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
  /**
     * @minLength 5
     * @maxLength 300
     */
  address: string;
  serviceType: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  priority: QuoteInputPriority;
  imageUrls?: string[];
}

export type QuoteUpdateStatus = typeof QuoteUpdateStatus[keyof typeof QuoteUpdateStatus];


export const QuoteUpdateStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface QuoteUpdate {
  status?: QuoteUpdateStatus;
  adminNotes?: string;
}

export interface QuoteList {
  quotes: Quote[];
  total: number;
  page: number;
  limit: number;
}

export interface QuoteStats {
  total: number;
  pending: number;
  reviewing: number;
  in_progress: number;
  completed: number;
  todayCount: number;
  weekCount: number;
}

export type ServiceCategory = typeof ServiceCategory[keyof typeof ServiceCategory];


export const ServiceCategory = {
  residential: 'residential',
  commercial: 'commercial',
  industrial: 'industrial',
} as const;

export interface Service {
  id: number;
  name: string;
  description: string;
  category: ServiceCategory;
  icon: string;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface AdminSession {
  id: number;
  email: string;
  name: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}

export type ListQuotesParams = {
status?: ListQuotesStatus;
page?: number;
limit?: number;
};

export type ListQuotesStatus = typeof ListQuotesStatus[keyof typeof ListQuotesStatus];


export const ListQuotesStatus = {
  pending: 'pending',
  reviewing: 'reviewing',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;