# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */
import type { QuotePriority } from './quotePriority';
import type { QuoteStatus } from './quoteStatus';

export interface Quote {
  id: number;
  name: string;
  email: string;
  phone: string;
  address: string;
  serviceType: string;
  description: string;
  priority: QuotePriority;
  status: QuoteStatus;
  /** @nullable */
  adminNotes?: string | null;
  imageUrls?: string[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}