# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteInputPriority } from './quoteInputPriority';

export interface QuoteInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
  /**
     * @minLength 5
     * @maxLength 300
     */
  address: string;
  serviceType: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  priority: QuoteInputPriority;
  imageUrls?: string[];
}