# lib/api-zod/src/generated/types/quoteList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */
import type { Quote } from './quote';

export interface QuoteList {
  quotes: Quote[];
  total: number;
  page: number;
  limit: number;
}