# lib/api-zod/src/generated/types/quoteStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteStats {
  total: number;
  pending: number;
  reviewing: number;
  in_progress: number;
  completed: number;
  todayCount: number;
  weekCount: number;
}