# lib/api-zod/src/generated/types/quoteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteUpdateStatus } from './quoteUpdateStatus';

export interface QuoteUpdate {
  status?: QuoteUpdateStatus;
  adminNotes?: string;
}