# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pozo Home & Business API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from './serviceCategory';

export interface Service {
  id: number;
  name: string;
  description: string;
  category: ServiceCategory;
  icon: string;
}